requirements.txt
aws-cdk.core>=1.134.0
aws-cdk.assertions>=1.134.0
aws_cdk.aws_lambda>=1.134.0
